fix/Gallery folder restructure
Created a folder for gallery components and a separate .css file for styling
