+++ src/app/(frontend)/components/Gallery/Gallery.css
@@ -1,0 +1,34 @@
+.gallery-grid {
+  column-count: 3;
+  column-gap: 0.5rem;
+  margin-top: 1rem;
+  margin-bottom: 5rem;
+}
+
+.gallery-grid-item {
+  margin-bottom: 0.2rem;
+  break-inside: avoid;
+}
+
+/*Gallery Grid Breakpoint*/
+@media (min-width: 1024px) {
+  .gallery-grid {
+    column-count: 6;
+  }
+}
+
+.albums-header-section h2,
+.albums-header-section h3 {
+  margin-top: 0;
+}
+
+.albums-header-section {
+  display: flex;
+  flex-direction: column;
+  gap: 30px;
+  margin-bottom: 90px;
+}
+
+.albums-header-row h3 {
+  font-style: italic;
+}
