+++ src/app/(frontend)/components/styles.css
@@ -18,40 +18,7 @@ body {
   margin: auto;
 }
 
-.gallery-grid {
-  column-count: 3;
-  column-gap: 0.5rem;
-  margin-top: 1rem;
-  margin-bottom: 5rem;
-}
 
-.gallery-grid-item {
-  margin-bottom: 0.2rem;
-  break-inside: avoid;
-}
-
-/*Gallery Grid Breakpoint*/
-@media (min-width: 1024px) {
-  .gallery-grid {
-    column-count: 6;
-  }
-}
-
-.albums-header-section h2,
-.albums-header-section h3 {
-  margin-top: 0;
-}
-
-.albums-header-section {
-  display: flex;
-  flex-direction: column;
-  gap: 30px;
-  margin-bottom: 90px;
-}
-
-.albums-header-row h3 {
-  font-style: italic;
-}
 .nav-button {
   justify-items: center;
   display: flex;
